refactor: handle page data hmr in the same file
